# Avoid repeated os.SetEnv calls in container init.