In a Kubernetes cluster with a large number of services (>5000, causing a large amount of environment variables for the container) with guaranteed pods, the `runc init` process can take a long time to run. This is caused by two issues:

1. `runc init` calls `os.SetEnv` for every environment variable in the bundle, which can be expensive due to the underlying mutex locking and syscall logic. 
2. The cgroup is limited before the `runc init` process runs. This is outside the perview of runc I believe.

This PR targets the former by removing the `os.SetEnv` calls, which causes the `runc init` process to run much faster in cases where there are a large number of configured environment variables in the bundle.