@kolyshkin 2024-06-15
This looks interesting. Would be nice to add a benchmark.