@kolyshkin 2024-06-15
Did some quick benchmark with 5000 env vars, comparing `os.Setenv` with passing env via cmd.Env. The latter is indeed faster!

```
name     old time/op    new time/op    delta
Exec-20    4.71ms ±33%    2.08ms ±13%  -55.88%  (p=0.000 n=10+9)

name     old alloc/op   new alloc/op   delta
Exec-20    1.43MB ± 0%    0.46MB ± 0%  -67.83%  (p=0.000 n=10+9)

name     old allocs/op  new allocs/op  delta
Exec-20     5.19k ± 0%     0.03k ± 0%  -99.46%  (p=0.000 n=10+10)
```

Will carry this one.